Checks for null vs nil, as they behave differently
It seems some values in `optionsOrNil` are coming out as null. You can't send messages to null, so, e.g. the `longLongValue` call crashes, vs returning nil when called on nil. I'm not sure if the behaviour of this has changes or the behaviour of `optionsOrNil` has changed to start doing this, or if it's always been the case, but only now exposed.